Move latest incident component to landing folder
