+++ site/gatsby-site/src/components/landing/LatestIncidentReport.js
@@ -3,7 +3,7 @@ import { StaticQuery, graphql, Link } from 'gatsby';
 import styled from 'styled-components';
 import md5 from 'md5';
 import { format } from 'date-fns';
-import { Image } from '../utils/cloudinary';
+import { Image } from '../../utils/cloudinary';
 import { fill } from '@cloudinary/base/actions/resize';
 import { Card, Col, Row } from 'react-bootstrap';
